some optimization in giem2g. change some int to long in extreme

+++ EventArgs/DotProductRequestEventArgs.cs
@@ -4,14 +4,14 @@ namespace Extreme.Fgmres
 {
     public unsafe class DotProductRequestEventArgs : System.EventArgs
     {
-        public int Length { get; }
+		public long Length { get; }
         public int NumberOfDotProducts { get; }
         public Complex* X { get; }
         public Complex* Y { get; }
         public Complex* Result { get; }
 
 
-        public DotProductRequestEventArgs(int length, Complex* x, Complex* y, Complex* result, int numberOfDotProducts)
+        public DotProductRequestEventArgs(long length, Complex* x, Complex* y, Complex* result, int numberOfDotProducts)
         {
             Length = length;
             X = x;
